OJ-2209: Set maximum preview stack name length limit
The common stack name is used in lambda names in the format `<common-stack-name>-<lambda-name-suffix>`. CloudFormation allows a maximum stack name length of 128 characters but the lambda service allows a maximum lambda name length of 64 characters. Therefore, the stacks we deploy must have a maximum length of `64 - <longest lambda name suffix> - 1` (1 char for the hyphen).

The longest fixed lambda name prefix we use is `AuditEventConsumerFunction` (26 chars), therefore, the stack name must not exceed 37 characters to meet this arbitrary constraint.