LIME-1071 Update to CRI Lib 3.0.1 and align with service changes.
AWS SDK 2.20.162 -> 2.26.16
New Aws Crt Http Client aligned with AWS SDK 2.26.16
AWS Lambda Events 3.11.0 -> 3.11.6
AWS Lambda Powertools 1.12.3 -> 1.18.0
Nimbusds Oauth 11.2 -> 11.4
Jackson 2.13.1 -> 2.15.2

Removes unused dependencies in the Jackson configuration.

Remove versions specified for the main Jackson dependencies and defer to the ones pull in from software.amazon.awssdk:bom.
There is a breakage in a later jackson version, which crashes in the sdk classes which expect the older versions.
Note : jackson-datatype-jsr310 and jackson-datatype-jdk8 are not in the aws sdk and are custom dependencies these version have been set at the aws pom versions to

Corrected deploy-branch being unaware of changes, as it was restoring and deploying the first cached build on the branch