OJ-2651: Implement typescript token pact test
The express application that exposes the /token endpoint needs
to be more generic as it would be used by all CRI's that consumes
the typescript token endpoint

Implement access-token-router which mocks out parameters using
MockSSMProvider that allows the getParametersByName to return values
just as the actual SSMProvider would but without making aws calls

MockDynamoDBDocument that allows mocking dynamo send and query operations
with the expected structure holding `sessionItemInDbBeforeTokenRequest`
for both outputs of send and query respectively.

The componentId needs to be overridden depending on the CRI, this is achieved
by using requestFilters in the pact verifier. The override is set as an
header value that is retrieve is the request middleware and used in
access token router.

Co-authored-by: Mike Beeby<mike.beeby@digital.cabinet-office.gov.uk>