OJ-2978: Create mock-jwks handler
Currently, the core-stub signing key is a is
an elliptic-curve cryptography key with curve P-256 (ECC_NIST_P256)
that was generated outside of KMS

This is so the core-stub has access to the private-key
which can be then be used to sign the JAR. This key does not belong
to CRI's so typically it not ideal as a KMS key in the CRI. Therefore,
core-stubs for CRI's use cases are signed this way.

This lambda generates the a new key-pair an and exposes as a well-known endpoint
in addition to the existing key already in uses by existing core-stubs.

The old Public signing verification keys will be used to verify the JAR was sent
by the headless-core-stub signed with it's corresponding private signing key in this case
in position zero of the jwks array retrieved from the endpoint.

Potentially, there can be two signing keys in the .well-known/jwks.json
for instance during rotation. and request can potentially be signed by
any of the two keys. The second key is a valid key that is capable of signing
a JAR as well as it public signing key also valid for verifying JAR requests.