OJ-3080: Add jwt_verification_failed metric to TS session and token
Note: There seems to be an issue with metrics not being published when errors are thrown by the validateJwtMiddleware and caught by the errorMiddleware. I have added metrics.publishStoredMetrics(); in the errorMiddleware to make sure the new metric is published. This can/should be looked at in more detail if needed in the future.