OJ-3183: Enable key rotation and fallback flags
Enabled key rotation and fallback flags for Check HMRC
Existing KMS key can now be accessed throw wellknown jwks endpoint

However this key is un aliased, so during decryption the process
will fallback to the default key existing if the un-aliased key is not found.
which is case as the rotated key has been overwritten. by the publish lambda

https://github.com/govuk-one-login/ipv-cri-common-infrastructure/pull/124