OJ-3183: slim down start integration tests
Since turning the key rotation flags in the stub, which means the
JAR requests will be made by using the JWKs endpoint of the CRI, which
is located in the CRI account, the decryption and verify aspects of the
integration tests are no longer possible. Originally, the tests were using
the core-infrastructure keys are deployed also in the stub, this why they
were originally able to work. The Decryption verifier aspect is a concern
of the CRI, perhaps we can have a CRI stub to fulfil this role in future.
For Decryption and verification will be moved to the Check HMRC CRI.